Update patch set 5
Patch Set 5:

Hi Martin

please check

http://i.imgur.com/cS4Z6.png

http://i.imgur.com/ZnWpS.png

before step forward I want confirm with you, are these just
what you want in below 2 ideas? :)
 

"For # 2 I would suggest an icon instead of the "All" word, but I would suggest an icon which is related to a note rather than the one proposed since I think we use that one for editing, and we aren't really editing the file?"

"For #4 I suggest that instead of changing the color of the comments, that the green header color be extended down around the comment (so that the column where the line numbers are becomes greeen around a whole file comment)."

Thanks a lot!

/Bruce

Patch-set: 5