Update patch set 2
Patch Set 2:

I am undecided on this change. It is user friendly to include the init command in the error message but it is also something that can get outdated if we change something in the Init command options/arguments
and forget to cross check this place.

Patch-set: 2
Reviewer: Gerrit User 1003873 <1003873@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=0
Label: Verified=0