Update patch set 6
Patch Set 6:

I can see where that would be annoying. I'm not sure what the right fix is. Note that (AFAICT) even without this change there would be nag emails on every server restart, since prior to this change we were still rescanning for unmerged changes on startup.

Do you think it would be feasible to add logic to MergeOp to specifically avoid this? Basically, somehow track the status of the last merge attempt, and don't generate emails if both the current and last attempts failed due to missing dependencies.

As a workaround in the meantime, just set changeMerge.checkFrequency = 0 to revert to the old behavior.

Patch-set: 6