Update patch set 6
Patch Set 6:

Thanks for the fast reply. First of all this is not very urgent for us, since at the moment we are running Gerrit 2.5 which doesn't contain this change, but we would like to upgrade to master quite soon and then it would become an issue.

It's correct that before this change there was the same problem on every server startup. But since our servers are seldomly restarted this was kind of acceptable.

Yes, it would be good to adapt MergeOp in the way you describe. I haven't looked into the details yet, but I think this should be possible.

Patch-set: 6