Update patch set 4
Patch Set 4:

This change introduced a problem.

When you have an open change in review and you bypass code review pushing another commit that has the same Change-Id (to the same project / destination branch), then the change in review should be automatically closed, means its state should be set to 'merged'. This was working fine until this change... :(

Patch-set: 4