Update patch set 1
Patch Set 1:

Hmm, it's probably because I made the ProjectScreen inherit directly from Screen instead of MenuScreen.  You wouldn't think that this would happen, but perhaps certain widgets on those screens were not designed right to start with?  I will see what I can do in the morning, it might be simplest to just re-inherit from the MenuScreen, but that will give you a bar down the left side of the screen (which I actually like since it gives back some context about the screens you are on, but I suspect it would look weird to most people).

Patch-set: 1
Label: Code-Review=0