Update patch set 1
Patch Set 1:

An empty bar on the left side of the screen sounds like a strange UI concept, although I agree that it would probably not look bad. Actually I think that the project screens look a bit unaligned on the left side of the screen, because the top level menu doesn't have a background color with the new CSS anymore it looks like there is a space on the left side of the menu which is missing for the controls below.

  http://i.imgur.com/XV0Ky.png

Patch-set: 1