Update patch set 2
Patch Set 2:

The new CommitMessageEditedSender and existing RebasedPatchSetSender are already derived classes of ReplacePatchSetSender that override the formatChange method.

Allowing them to optionally provide their own template or use a default makes it more complex, not simpler.

I agree that it would be useful to be able to do that.  Server admins might find it useful to be able to reuse the same template.  However I think that work, if it indeed is needed, should be done in a more wide scoped refactoring of the mail sending functionality, and is not something that should block this change from going in.  As it is now, there is no mail sent when the commit message is edited, and we need to fix that.

BTW Shawn had mentioned [1] replacing velocity, so might not be worth putting effort into refactoring yet.

[1] https://gerrit-review.googlesource.com/#/c/38921/1//COMMIT_MSG

Patch-set: 2