Update patch set 1
Patch Set 1:

I'm planning to add a new ExtensionPoint,  allowing plugins to contribute new download links. If the name of the downloadlink  matches one of the default ones, it will just override it. I think this is a fair way to attack these kind of download link customizations.

Patch-set: 1
Label: Code-Review=0