Update patch set 2
Patch Set 2:

It could use the same template, but then it would be inconsistent with the rebase feature.

I'd like to get this submitted, and then start work on the changes that have been suggested.  I.e. make rebase and edit commit message both use the same template as new patchset, but with the possibility for the admin to supply a custom template.

Patch-set: 2