Update patch set 3
Patch Set 3: Code-Review+2

(1 comment)

As a reader, I have been trying to understand all the different methods now proposed and this one does feel the easiest to follow (it even allows the for loop to be put inside the try block if desired, or for the exception to be ignored and thrown by your function) , and the callback is the hardest.  If Gustaf's stuff can be easily made to use this I, I would recommend it.

Patch-set: 3
Label: Code-Review=+2