Update patch set 1
Patch Set 1:

Could you provide an example of such a dashboard which you think makes sense to display on the change page and a screenshot for it?

What about the dependencies panel on the same screen in the light of this feature? Is it becoming just another dashboard? In other words, does it make
sense to remove the hardcoded dependencies panel and leave it to site/project admins to include it on their change screen as a dashboard?

Patch-set: 1