Update patch set 1
Patch Set 1:

I implemented adding and deleting group members as batch operations because I thought it's good to

 a) save roundtrips between client and server
 b) save roundtrips between server and database
 c) have it as atomic operations.

But I see now that having a DELETE request with a body is bad and we should get rid of it (for the reasons that Shawn mentioned).

I think the most rest-ful way to do batches is HTTP pipelining [1], but since this wasn't giving me b) and c) I wasn't looking into this. Is this the kind of pipelinig that we do now?

For deleting members I actually think neither performance nor atomicy is very crucial, but we should decide on a solution that will also fit for other batch delete operation that we might want to implement.

I agree with Shawn, that being able to identify delete operation in the logs immediately by seeing DELETE as method is a benefit, that's my d).

Maybe supporting DELETE /groups/{id}/members?q=user1&q=user2 is not so bad, as it gives us a), b), c) and d) until we hit the query limit. So the question is how likely this limit is hit? If it is hit, can we detect it beforehand and split the request?

I'm still undecided what is the best solution and I haven't looked at the code of this change yet (will do this later).

[1] http://en.wikipedia.org/wiki/HTTP_pipelining

Patch-set: 1