Update patch set 1
Patch Set 1:

So... we've actually been running this change in production for about a month and it's helped a lot.

As for the case you described Edwin, Gerrit doesn't touch the git repo (since it's already there), but it does update the database to mark the change as merged.

Patch-set: 1
Reviewer: Gerrit User 1002666 <1002666@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=0