Update patch set 1
Patch Set 1:

I do not think that the message is currently visible to the user in
stock gerrit. But it would help to migrate to a single consistent
default error message if a project could not be found. Currently, we
have
* gerrit-server/src/main/java/com/google/gerrit/server/args4j/ProjectControlHandler.java:68
 'X': is not a Gerrit project
* gerrit-sshd/src/main/java/com/google/gerrit/sshd/BaseCommand.java:431
  X no such project
  (I am not sure how to trigger this message)

For renaming projects, I'd add a third place, where we throw a
NoSuchProjectException.

So instead of introducing yet another place where we manually create a
default error message for NoSuchProjectException and thereby raise the
burden of keeping them consistent, it'd be good if
NoSuchProjectException would be consistent with NoSuchGroupException
and provides a descriptive, default problem description on its own.

I'll add a new patch set that harmonizes the current error messages.

Patch-set: 1
Reviewer: Gerrit User 1012155 <1012155@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=0