Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

As discussed in person I wonder if there's a better way to generalize this, but as it is it's not so bad.

Also, this doesn't match the semantics of our internal code review tool, where a sticky approval can't be removed even by negative votes. I'm guessing this is intentional and we prefer this for Gerrit.

Patch-set: 1
Reviewer: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+1