Update patch set 1
Patch Set 1:

I would start with only making the relative format available for the change table (ChangeTable2). This is the place where it makes most sense. I believe in other places, e.g. on the change screen, you would still like to see the absolute date.

Then it could maybe be a simple new boolean preference which is independent of the existing date/time preferences.

Patch-set: 1