Update patch set 2
Patch Set 2:

I dread the bugs this extra complexity might introduce.  Perhaps there is a clever CSS way to do this instead using sibling relationships to the preceding th element?  If the preceding th element is on an odd row, and we are on an odd row, then we should match?  Same for even rows.  However if there is an odd/even combo, make the color different?

Patch-set: 2