Update patch set 2
Patch Set 2: Code-Review-1

> there is no odd row immediately following an odd row.

But the headers rows can be designed to be of a different CSS class, you then will have a reference point in CSS for each row.  I agree that CSS is ugly at times, but I sure would hate to have to make the code ugly too.  I'm not convinced that this is impossible to do in CSS.

Patch-set: 2