Update patch set 9
Patch Set 9: Code-Review-1

(4 comments)

If there is a demand for this command now, I agree that we shouldn't block it. When I was starting this there was nobody wanting it urgently.

Would it make sense to have an own change for only adding the 'username' to AccountInfo? AccountInfo is used in quite some places and this change is probably worth to be discussed independently.

Patch-set: 9
Reviewer: Gerrit User 1003935 <1003935@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1