Update patch set 3
Patch Set 3:

Colby's argument is an OutOfMemoryError may have caused a background thread to corrupt its internal data structures in a way that cannot be recovered from. If this happens than the process may never recover, even if the GC manages to reclaim memory and can allocate again.

I agree with Jonathan, this is a very bad idea. The server shouldn't just up and go away.

Colby's argument is if we trapped OutOfMemoryError we may not have sufficient memory to send a reply to the client. So apparently we shouldn't even try. But in a perfect deployment environment there is some sort of job control watching the server and restarting it if it terminates unexpectedly. Gerrit doesn't provide that job control itself, instead we (mostly) assume the JVM stays up.

Patch-set: 3
Reviewer: Gerrit User 1001240 <1001240@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0