Update patch set 1
Patch Set 1:

Hi,

I suggest to not introduce the resizing search box in 2.6, when it is reverted in 2.7.

I've pushed this in order to not forget it, but I'm unable to build it currently, as maven cannot resolve plugin dependencies and I'm on travel atm.

Hope you still like it :-)

Steffen

Patch-set: 1
Reviewer: Gerrit User 1011597 <1011597@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0