Update patch set 3
Patch Set 3:

Any ideas on how we can resolve this for 2.7?  It would be useful to wrap this one up for 2.7, then if we need to redesign how the LDAP backend does its group visibility (and group visibility in general) for 2.8+, I'll be willing to lend a hand with that.  Shawn, your insight into how the current Google internal backends work at a high level would be appreciated, too, since without that kind of knowledge, we're at something of a disadvantage trying to make this backend match it.

Patch-set: 3