Update patch set 5
Patch Set 5:

Currently we can't run the acceptance tests in parallel, they bind to the same hardcoded port numbers in the tests. The tests have to be rewritten to use random port selection to prevent collisions.

Buck threads tests at the java_test() level of parallelism, their threading is based on the build system threading. So we would also have to decompose the acceptance tests target into smaller targets. Which I am happy to do if the port thing is fixed.

20 minutes may be a reasonable time for these tests to run. They might need even more. I usually run them and go to lunch or whatever. Maybe there is something else wrong in there.

Patch-set: 5