Update patch set 8
Patch Set 8:

When duplicates exist we throw InvalidPlugin exception while trying to load further plugins with the same name. Given that delete-project plugin include that line in MANIFEST file:

 Gerrit-PluginName: deleteproject

and we put two files with this MANIFEST in $site/plugins directory:

 * delete-project-2.7-SNAPSHOT.jar
 * delete-project-2.8-SNAPSHOT.jar

then we get the following exception when loading the second plugin:

 WARN PluginLoader : Cannot load plugin delete-project-2.8-SNAPSHOT
 InvalidPluginException: Duplicate plugin name: deleteproject. The same plugin was already loaded from: delete-project-2.7-SNAPSHOT.jar

I think this is very logical and expected behavior in this a case and conform with the principle of least surprise.

Patch-set: 8
Label: Verified=0