Update patch set 10
Patch Set 10:

That might be useful, but ideally I'd like *all* the htmls to be there so I can just launch the index file and browse around, which is useful to be able to do when changing several files at the same time.

If they're always going to be available in that temp folder it's more or less OK as it is.

Patch-set: 10