Update patch set 6
Patch Set 6:

This slows down reindexing with Lucene, probably because there is more waiting on futures and we're not getting as much parallelism. I plan to look into this, and I consider it blocking for turning on index support by default, but for now I want to submit because the next change in the series fixes several bugs in the live googlesource.com implementation.

Patch-set: 6