Update patch set 6
Patch Set 6: Code-Review+1

1. The same user is listed more than once 
We have corrected the code to show the user only once (when the change is merged and no reviewer is added).

2. Users are listed even though they are not actually a reviewer 
This change is suppose to show all the reviewers who are added to the change irrespective of whether they have reviewed or voted on that change.
In case you are referring to something else please specify so we can correct it.

3. 'currentReviewers' is plural but is not a list. It is repeated multiple times, one for each reviewer. 
Here it is following the same pattern as in --all-approvals for each approval it shows the tag 'approvals'.
Hence, for each reviewer it shows the tag 'currentReviewers', however if the query is viewed in JSON format it shows the tag 'currentReviewers' only once.


4. Empty 'currentReviewers' appears sometimes
Empty 'currentReviewers' appears only when there is no reviewer added and change is not merged.
However, we checked with other options like --submit-records and --dependencies and they also show an empty bracket.
Hence, we followed the same pattern.


For example (case 1, 2, 3) on a change that does not have any current reviewers and is in status "merged" because I force-pushed it:

The user who is merging the code is also a reviewer (as per the requirement) hence it shows the name of that user.
We have corrected the code to show the name only once.

Patch-set: 6
Label: Code-Review=+1