Update patch set 1
Patch Set 1:

This somehow was based on the 'merge stable-2.8' commit on master, and has caused a couple of unwanted commits from master to be included in stable-2.8

Sorry, I should have noticed this when reviewing it.  I will upload a revert to get rid of the unwanted changes.

Patch-set: 1