Update patch set 2
Patch Set 2:

Regarding the height, I don't think so. Its about the same height as the current panel. Most commit messages should have at least one additional paragraph in the body supporting why the thing that is happening in the subject line is a good idea. Most real changes have a handful of reviewers and votes on them, which increases the height of the info table.

The message I used here was quite intentional. I tried to keep it somewhat short, but the right margin is exactly at 80 columns, except for the URL which obviously overflows. I used this commit in my development environment to see how the new layout handles it. We actually lose the footer lines, they aren't visible unless you scroll or expand. But it is a heck of a lot better than what we have now.

Most commits in Gerrit are multiple lines in the commit message. I think this is true for most Git projects in general. So I'm happy with the height.

Patch-set: 2
Reviewer: Gerrit User 1001240 <1001240@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0