Update patch set 2
Patch Set 2: Code-Review-1

I really don't want to put this into 2.6. We only support the last released version, which is 2.7. And we are trying to ship 2.8. But we keep getting stuff on 2.6, and there is no 2.6.2 release in sight. I don't see this as a critical bug fix either, its a new feature that the logs are configurable.

Patch-set: 2
Reviewer: Gerrit User 1001240 <1001240@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1