Update patch set 12
Patch Set 12:

I may have spoken too soon.  It certainly resolves the stack overflow the first time around, but the second time you try to update preferences, it appears to come back.  Anyway, this is a bit cleaner copy of what I was thinking of, including splitting out some resources so we're not relying on the help panel's styles, and also splitting the labels into a separate resource as well.

Patch-set: 12
Label: Verified=0