Update patch set 1
Patch Set 1:

This is a big imrovement. One small problem: when i edit a draft, then scroll down so that the box is out of focus, then type "u", then on the CS2 draft:1 is not appear in the file table for edited file.

Only if i reenter the CM3 on the same file again and return back to the CS2 (without even to edit it again), the draft:1 appears.

Patch-set: 1
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=0