Update patch set 1
Patch Set 1:

I wouldn't mind, except it would still have the problem that msys is weird (and I hate that, but it's also 95% of the userbase I support).  Honestly, If this isn't preferred, I'd rather just revert Ie4fcd0a73bd1cee12a052f38541f83aa8821c886 and call it a day.  Also, the 2.8 version looks like less than a trivial modification, so would it really be something we want to copy back to 2.6?

Patch-set: 1