Update patch set 1
Patch Set 1:

I agree; there shouldn't be two different flavors of the command we present to users between the two versions, but the only reason I prefer this is the oddness that Git for Windows presents -- if it weren't for that, I'd be all for just backporting the change made to 2.8.

I have tried what's used in 2.8/master, and the problem is the command substitution triggers an odd bug: the scp commandline doesn't get parsed right if the path returned by "git rev-parse --git-dir" is included directly... but if it's substituted via a variable, everything works fine.  Basically, this is only when you're dealing with a path like "D:/Projects/gerrit/.git/modules/plugins/replication-plugin/" -- I'm not entirely sure I like including a workaround for a MSYS bug here, but since it's simple (store the value in a variable first) and it works, that's why I proposed using that instead, ugly as it may be.

Patch-set: 1