Update patch set 4
Patch Set 4:

The idea behind one "CredentialVerifier per specific Credentials type" is that it will be too hard to support every arbitrary combination and use case initially. If you need something more complicated, you should write your own CredentialVerifier that handles the Credentials in the order you expect based on the incoming user. The goal behind this is to provide the basic workflow that makes sense and is plugable, but not solve every auth flow in the core.

Patch-set: 4