Update patch set 1
Patch Set 1:

I am consciously resisting premature generalization. If we end up having three or four or ten different tombstones down the line I'm willing to reconsider then.

Keep in mind the human-understandability of the format. We are not CRUDding records in a database; we are making semantic updates to a change object. It is entirely reasonable IMHO for the commit generated by clicking the X on a reviewer chip to have the text "Reviewer: XYZ=REMOVED". (Even better if the account ID were less opaque of course :)

Patch-set: 1
Label: Verified=0