Update patch set 1
Patch Set 1:

Hey Dave, can you tell me if this is Diffy's fault or if this is correct?  Should there be a change to a BUCK file or anything?

Diffy runs a build without the change and keeps a copy of ~/.gerritcodereview/buck-cache so that a second build (with the patchset) can be done without network access.  Would these needed objects only get thrown in buck-cache if they are imported?

Patch-set: 1
Reviewer: Gerrit User 1006257 <1006257@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0