Update patch set 1
Patch Set 1:

I have been running 2.8 with the H2 DB jar replaced with 1.3.174 for a couple of days.

I haven't noticed any side effects - but as we hit this somewhere between twice a week and once a fortnight its too early to say if this has cured the problem.

Patch-set: 1