Update patch set 3
Patch Set 3:

> > Thanks, I wasn't aware of this. Unfortunately I'm still getting
 > > OutOfMemoryError even if I have .buckjavargs with these settings.
 > > I've added -XX:+HeapDumpOnOutOfMemoryError but I'm not getting a
 > > heap dump which makes me think that this file is ignored on my
 > > system... :(
 > 
 > You put this .buckjavargs file in gerrit top directory, right?

Yes.

 > In ~/projects/buck/bin/buck_common they are putting both values on
 > the command line:
 > 
 > BUCK_JAVA_ARGS="\
 > -XX:MaxPermSize=256m \
 > -Xmx1000m \
 > [...]
 > ${BUCK_PROJECT_JAVA_ARGS} \
 > ${BUCK_EXTRA_JAVA_ARGS}"
 > 
 > May be delete the first two lines? You can also change the first
 > line of buck script to be:
 > 
 > #!/bin/bash -x
 > 
 > to see verbose output.

I inspected this a bit closer. While the test is being executed I see 2 processes on my system:

1. one for com.facebook.buck.cli.Main test //gerrit-acceptance-tests/src/test/java/com/google/gerrit/acceptance/server/project:LabelTypeIT

2. one for com.google.gerrit.acceptance.server.project.LabelTypeIT

The first process (1.) is executed with the memory settings that I have configured in .buckjavaargs and the setting that I can see in buck/bin/buck_common. Making a heap dump of this process shows that it's actually consuming less than 9MB heap.

The second process (2.) is executed only with '-Xmx128m'. I assume this is the process where I'm getting an OutOfMemoryError. Trying to make a heap dump of this process fails and I don't know how I can change the memory settings for this process.

Patch-set: 3