Update patch set 13
Patch Set 13: Code-Review-1 Verified-1

With the current implementation this is not safe against XSS attacks. Dave raised the question about XSS on [1]. I though we are safe since we suppress all HTML that is contained in the Markdown file when we generate the project documentation HTML [2]. However as explained in [3] this is not sufficient since XSS can also be done with Markdown links:

  [some text](javascript:alert('xss'))

becomes

  <a href="javascript:alert('xss')">some text</a>

after the generation. 

To be safe we need to filter the resulting HTML for XSS. Does anyone know a Java library for XSS filtering that has a suitable license? 

 [1] https://gerrit-review.googlesource.com/53919
 [2] https://gerrit-review.googlesource.com/#/c/53010/12/gerrit-server/src/main/java/com/google/gerrit/server/documentation/MarkdownFormatter.java
 [3] http://michelf.ca/blog/2010/markdown-and-xss/

Patch-set: 13
Label: Code-Review=-1
Label: Verified=-1