Update patch set 14
Patch Set 14: Verified+1 Code-Review+1

> However as explained in [3] this is not sufficient since XSS can  
 > also be done with Markdown links:
 > 
 > [some text](javascript:alert('xss'))
 > 
 > becomes
 > 
 > <a href="javascript:alert('xss')">some text</a>
 > 
 > after the generation.

This was a false alert. My noXSS test was just wrong. 

Actually 

  [some text](javascript:alert('xss'))

turns into

  <p>[some text](javascript:alert(‘xss’))</p>

So this should be fine.

Still having an additional XSS filter would be nice. I'm currently checking whether we could use AntiSamy [1] for this. It has a BSD licence, is this okay for us?

Anyway, if using AntiSamy works out it will be an own change. I'm currently fighting with getting all the needed dependencies in place...

[1] https://www.owasp.org/index.php/Category:OWASP_AntiSamy_Project

Patch-set: 14
Label: Code-Review=+1
Label: Verified=+1