Update patch set 1
Patch Set 1: Code-Review-1

I understand that this is most likely to show up with status:closed because that set grows without bound, but theoretically this applies to any query. A slightly more general solution would be to cap the limit the LimitPredicate constructor (but this still wouldn't handle queries that have no such predicate).

(It also feels like it belongs at the gwtorm layer because it's SQL engine specific, but I'm not sure offhand how to go about implementing that.)

Patch-set: 1
Label: Code-Review=-1