Update patch set 2
Patch Set 2:

Going ahead and publishing this, since I plan to make a release WAR at this version available on gerrit-releases.storage.googleapis.com. For technical reasons the Chrome builders need it to be on the public internet and not require authentication.

Not that it's useful for anybody else, I just don't want to publish something referring to a SHA-1 that isn't available.

Patch-set: 2