Update patch set 1
Patch Set 1:

The dashboard icon was added in stable-2.9; it should be removed there so we don't release a version with it, and then remove it in the next.

Also the dashboard icon resource is not used now, so should also be removed.

I've cherry-picked this to stable-2.9 with a follow-up change to remove the unused icon.

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0