Update patch set 3
Patch Set 3:

"Google Accounts" are also mentioned in "dev-design.txt" and "config-sso.txt". Do these docs need to be updated too?

Also this looks like something for which we should have a warning in the 2.10 release notes. Is it worth to add empty release notes with this warning already with this change, so that this warning is not forgotten later?

Patch-set: 3