Update patch set 4
Patch Set 4: Code-Review-1

Looking at this again I believe this change can cause some trouble.

Well, I do hate that my line breaks which I intentionally insert are simply ignored. The problem with this change is, that now long lines are not anymore wrapped automatically, but I'm forced to insert ALL line breaks manually. E.g. if I type a long paragraph as a comment, before this change it was nicely wrapped and displayed in several lines. With this change I will end up with one very long line and I have to scroll to read it, or I have to manually add a lot of line breaks, which is very annoying. Probably more annoying that having my own line breaks ignored. This is especially an issue when we look at existing comments. With this change they might suddently be rendered as a single line, making it very hard to read them :(

Patch-set: 4
Label: Code-Review=-1