Update patch set 17
Patch Set 17:

> > (1 comment)
 > >
 > > The Related Changes tab is not shown at all for a change having
 > > dependencies on other open changes in the same branch.
 > Thanks David.
 > I will check and get back to you.

I created 4 changes in the same branch which are dependent on each other and was able to see them in Related Changes Tab.
Here is the gitk result for all changes - http://imgur.com/NAj7b7O
Here is the screenshot of change showing changes in Related Changes Tab - http://imgur.com/TSpt37Q

Please let me know if this the case that you are referring to?
Thanks.

Patch-set: 17