Update patch set 4
Patch Set 4: Code-Review-1

I'm not really convinced that this is much of a simplification.

There are still a lot of errors being output with the `OutputStream err` so now we have a combination of that and `stderr.println`.

Also, I see there are some errors that are output to the `OutputStream out` which I assume goes to stdout.  Should that be cleaned up too?

Patch-set: 4
Label: Code-Review=-1