Update patch set 2
Patch Set 2:

I think the way this is implemented is very confusing.  It seems like the new link that's added should be a toggle link, but it doesn't act like it.  That along with the side-by-side link on the top navigation panel makes it unclear which one the user should select to change the view.  Would it be possible to just remove the top side-by-side completely and keep this new link but make it toggle between unified and side by side?  I think that would make it work a little better at least.

Patch-set: 2