Update patch set 1
Patch Set 1:

The core of the issue with the tests is that we are passing the output of rewrite.rewrite() to the IndexedChangeQuery constructor, when this is not how IndexedChangeQuerys are created in the normal rewrite code path.

One argument (ignoring for the moment issues with equals()) is that we should use QueryProcessor.parseQuery() so that the input to the index rewriter always matches something we could reasonably see in the wild.

But I think in this case it's probably ok to have IndexRewriteTest test a more simplified set of inputs, since the purpose of these small tests is to make sure we're doing the recursive algorithm correctly. Adding extra limit predicates where we don't need them gets in the way of this goal. And we already have separate end-to-end query tests (including for limit), so it's not like we're lacking coverage.

Overall, I think the checkNotNull is not the right way to go. Sorry for time wasted. I will fix up your PS1 and send it out when tests pass (which should be easy this time around :)

Patch-set: 1