Update patch set 13
Patch Set 13:

> you should also add information about this endpoint to dev-rest-api
 > documentation, either in this change on somewhere in the topic

We discussed it with David P. previously. We don't want to expose the RESTApi in the documentation at this point, since there is no information on how to enable/configure the NoteDB. It would probably just confuse the users. But I'll try to add this information in the commit message.

Patch-set: 13
Label: Verified=0