Update patch set 4
Patch Set 4:

Sven: I do not agree here. This is how reviewers can be added, and those can be removed. It has never been the source of confusion. It should also be very easy, using Davids HashtagValidatorListener, to write a plug-in which does not allow the removal of hashtags that exists in the commit message.

Patch-set: 4