Update patch set 7
Patch Set 7: Verified-1

Test pass in Eclipse, but fail on the command line.

testPushForMasterWithApprovals is also failing, both in Eclipse and on the command line.  This does not appear to be related to my changes, but I have not yet checked if it also fails on master.

Patch-set: 7
Label: Verified=-1