Update patch set 19
Patch Set 19:

IMOHO:
Hashtags in commit message should only be deleted by deleting them in commit message. That way the possibility to add them in commit message makes sense. I would like for this to be implemeted in core, but it's not vital for me.

Hashtags in commit message should not delete hashtags (not explicit nor implicit).

Patch-set: 19