Update patch set 21
Patch Set 21:

You can add a hashtag by having it in the commit message.
A hashtag that is in the commit message of the current patch set of the change, should never be removed.
By removing the hashtag from commit message in a new patch set you do not remove the hashtag, you only make it possible to remove it.

Patch-set: 21
Label: Verified=0