Update patch set 1
Patch Set 1:

I take it the main reasoning[1] is:

* LUCENE-5904: Fixed a corruption case...
* LUCENE-5958: Don't let exceptions during checkpoint corrupt the index...

http://lucene.apache.org/core/4_10_1/changes/Changes.html#v4.10.1

So yeah we should upgrade before the next release. But I still hate Lucene-only schema upgrades.

I think there's a >50% chance my mergeability series is going to involve a schema upgrade as well, so I'm inclined to wait for that.

Patch-set: 1
Label: Verified=0