Update patch set 9
Patch Set 9:

I'm not sure about Shawn's claim that we weren't going to upgrade...as of December 2013 we hadn't, but as of February 2012 we upgraded to 1.50 internally. It looks like we haven't picked up 1.51 yet, but if these methods were added in 1.49 we should be fine. Let me check.

Patch-set: 9