Update patch set 1
Patch Set 1:

> I suspect we will get pushback from users about this. It may be a
 > relatively rare use case that someone wants to edit just the commit
 > message and nothing else, but for when that does happen, the
 > workflow is very simple:
 > 
 > * Click "Edit Message"
 > * Edit the message
 > * Click Save
 > 
 > Now it's two more clicks:
 > 
 > * Click Edit
 > * Click Commit Message
 > * Edit the message
 > * Click Save
 > * Click Publish

Yes. I wasaware, this is going to be a controversial change. OTOH the big advantage to change everything in one place is that now the advantage of change edit workflow is also applies on changed commit message: changed commit message can be reviewed in SBS2 screen and for example by setting right margin line to 72 it can even be checked that commit message rules aren't violated. I added other reviewers to thing change to hear their opinion in this.

Patch-set: 1