Update patch set 4
Patch Set 4:

David P. - from the looks of it, the tag you linked earlier points to the same commit (94059021ce1f20dad838f9cc3670e29964e41173) -- so I'm not sure what changed? Still nothing on maven central, though.

As for an possible way to test this, we found when running Jenkins with a watchdog timeout, the frequency of encountering this bug increased (instead of being infrequent, it hit all instances we run after a weekend, and within a day of each other).  Like everyone else, we'd rather not "test in production," though... and seeing as we just managed to run a stock Gerrit build for once, I wouldn't be terribly excited to branch again. :) I suppose I could see about doing a custom build and reproducing this on a test environment...

Patch-set: 4
Reviewer: Gerrit User 1012612 <1012612@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0